Refactor webapi routes and init db module
