+++ services/favorites.py
@@ -5,7 +5,8 @@ from typing import Any, List
 
 from sqlalchemy import delete, select
 
-from database import get_session, init_db
+from database import get_session
+from initdb import init_db
 from models import Favorite
 from services import menu_catalog
