+++ services/promocodes.py
@@ -6,7 +6,8 @@ from typing import Any, Iterable
 
 from sqlalchemy import func, select
 
-from database import get_session, init_db
+from database import get_session
+from initdb import init_db
 from models import Order, PromoCode
